Removes light sensing from ocs.sh
The light sensor is dying.  It has no longer become a reliable way to
tell if the space is open.  This change removes the code associated with
detecting light and leaves behind the code for checking if the override
is set.  This will allow the code to be moved to a standalone
RaspberryPi with a touch screen.  When someone arrives at the space,
they simply hit the "Open" button on the touchscreen and all the various
social media and services are alerted to the spaces status.

In addition removing the light sensing code, this also rearranges the
code some to make it more readable.  I have moved the following
portions to their own functions:

- Opening
- Closing
- Clean up
- Logging

Finally, I removed some old commented out code.  If we need it again,
git stores all.