Updates logic to fix loop
After my last PR, the loop wasn't tracking the *change* of states which
is what we really care about.  If the space is currently open and it was
previously open, we don't care.  We only care when it is currently open
and previously closed.  The if statement has been changed to track these
changes now.